Avoid multiple database accesses on item update
Before this patch, for *every* user on an index page for a particular
resource type, an unnecessary query to the db was made after an item
update, even if the updated item was not currently in the list of
displayed items.

Same for Show views: `Backpex.LiveResource.Show.assign_item/1` was
called for all the currently active Show views of a given resource type,
even for those items that were not updated, resulting in unneeded
database queries.