fix `Backpex.Fields.Currency` breaking delete in `Backpex.Fields.InlineCRUD`
When a `Backpex.Fields.Currency` is used with
`Backpex.Fields.InlineCRUD`, the following bug happens. Delete one row
in the `InlineCRUD`, then make any change to the form: the deleted row
will suddenly reappear! This is because two values are sent to the
server: the old value sticks around because it is not updated due to
`phx-update="ignore"`.